add list component and move code from app.js to this file
